refactor - isolate generator
